src/main/java/kitchenpos/product/domain/Price.java
package kitchenpos.product.domain;

import java.math.BigDecimal;
import java.util.Objects;
import javax.persistence.Column;
import javax.persistence.Embeddable;

@Embeddable
public class Price {

    @Column(name = "price")
    private BigDecimal value;

    protected Price() {
    }

    private Price(final BigDecimal value) {
        validate(value);
        this.value = value;
    }

    public static Price from(final long value) {
        final BigDecimal price = BigDecimal.valueOf(value);
        return new Price(price);
    }

    private void validate(final BigDecimal value) {
        if (Objects.isNull(value)) {
            throw new IllegalArgumentException("가격은 null 일 수 없습니다.");
        }
        if (value.compareTo(BigDecimal.ZERO) < 0) {
            throw new IllegalArgumentException("가격은 음수일 수 없습니다.");
        }
    }

    public BigDecimal getValue() {
        return value;
    }

    @Override
    public boolean equals(final Object o) {
        if (this == o) {
            return true;
        }
        if (o == null || getClass() != o.getClass()) {
            return false;
        }
        final Price price1 = (Price) o;
        return Objects.equals(value, price1.value);
    }

    @Override
    public int hashCode() {
        return Objects.hash(value);
    }

    @Override
    public String toString() {
        return "Price{" +
                "price=" + value +
                '}';
    }
}
